# backend/openapi/specs/data-cleaning.yaml
openapi: "3.0.1"
info:
  title: "API Documentation"
  version: "1.0"
paths:
  /cleaning/templates:
    get:
      summary: "cleaningTemplatesGet"
      description: "cleaningTemplatesGet"
      parameters:
      - name: "page"
        in: "query"
        description: ""
        required: false
        schema:
          type: "integer"
          nullable: false
      - name: "size"
        in: "query"
        description: ""
        required: false
        schema:
          type: "integer"
          nullable: false
      - name: "keywords"
        in: "query"
        description: ""
        required: false
        schema:
          type: "string"
          nullable: false
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "string"
                    description: "状态码"
                  message:
                    type: "string"
                    description: "消息"
                  data:
                    type: "object"
                    properties:
                      page:
                        type: "integer"
                        description: ""
                        format: "int64"
                      size:
                        type: "integer"
                        description: ""
                        format: "int64"
                      totalElements:
                        type: "integer"
                        description: ""
                        format: "int64"
                      totalPages:
                        type: "integer"
                        description: ""
                        format: "int64"
                      content:
                        type: "array"
                        description: ""
                        items:
                          type: "object"
                          properties:
                            id:
                              type: "string"
                              description: ""
                            name:
                              type: "string"
                              description: ""
                            description:
                              type: "string"
                              description: ""
                            instance:
                              type: "array"
                              description: ""
                              items:
                                type: "object"
                                properties:
                                  id:
                                    type: "string"
                                    description: ""
                                  name:
                                    type: "string"
                                    description: ""
                                  description:
                                    type: "string"
                                    description: ""
                                  version:
                                    type: "string"
                                    description: ""
                                  inputs:
                                    type: "string"
                                    description: ""
                                  outputs:
                                    type: "string"
                                    description: ""
                                  runtime:
                                    type: "string"
                                    description: ""
                                  settings:
                                    type: "string"
                                    description: ""
                                  isStar:
                                    type: "boolean"
                                    description: ""
                                  createdAt:
                                    type: "string"
                                    description: ""
                                  updatedAt:
                                    type: "string"
                                    description: ""
                                description: "OperatorResponse"
                              default: "new ArrayList<>()"
                            createdAt:
                              type: "string"
                              description: ""
                            updatedAt:
                              type: "string"
                              description: ""
                          description: "CleaningTemplate"
                    description: "数据"
    post:
      summary: "cleaningTemplatesPost"
      description: "cleaningTemplatesPost"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                name:
                  type: "string"
                  description: ""
                description:
                  type: "string"
                  description: ""
                instance:
                  type: "array"
                  description: ""
                  items:
                    type: "object"
                    properties:
                      id:
                        type: "string"
                        description: ""
                      overrides:
                        type: "object"
                        properties:
                          key:
                            type: "object"
                            properties: {}
                        description: ""
                        default: "new HashMap<>()"
                    description: "OperatorInstance"
                  default: "new ArrayList<>()"
              description: ""
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "string"
                    description: "状态码"
                  message:
                    type: "string"
                    description: "消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "string"
                        description: ""
                      name:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      instance:
                        type: "array"
                        description: ""
                        items:
                          type: "object"
                          properties:
                            id:
                              type: "string"
                              description: ""
                            name:
                              type: "string"
                              description: ""
                            description:
                              type: "string"
                              description: ""
                            version:
                              type: "string"
                              description: ""
                            inputs:
                              type: "string"
                              description: ""
                            outputs:
                              type: "string"
                              description: ""
                            runtime:
                              type: "string"
                              description: ""
                            settings:
                              type: "string"
                              description: ""
                            isStar:
                              type: "boolean"
                              description: ""
                            createdAt:
                              type: "string"
                              description: ""
                            updatedAt:
                              type: "string"
                              description: ""
                          description: "OperatorResponse"
                        default: "new ArrayList<>()"
                      createdAt:
                        type: "string"
                        description: ""
                      updatedAt:
                        type: "string"
                        description: ""
                    description: "数据"
  /cleaning/templates/{templateId}:
    get:
      summary: "cleaningTemplatesTemplateIdGet"
      description: "cleaningTemplatesTemplateIdGet"
      parameters:
      - name: "templateId"
        in: "path"
        description: ""
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "string"
                    description: "状态码"
                  message:
                    type: "string"
                    description: "消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "string"
                        description: ""
                      name:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      instance:
                        type: "array"
                        description: ""
                        items:
                          type: "object"
                          properties:
                            id:
                              type: "string"
                              description: ""
                            name:
                              type: "string"
                              description: ""
                            description:
                              type: "string"
                              description: ""
                            version:
                              type: "string"
                              description: ""
                            inputs:
                              type: "string"
                              description: ""
                            outputs:
                              type: "string"
                              description: ""
                            runtime:
                              type: "string"
                              description: ""
                            settings:
                              type: "string"
                              description: ""
                            isStar:
                              type: "boolean"
                              description: ""
                            createdAt:
                              type: "string"
                              description: ""
                            updatedAt:
                              type: "string"
                              description: ""
                          description: "OperatorResponse"
                        default: "new ArrayList<>()"
                      createdAt:
                        type: "string"
                        description: ""
                      updatedAt:
                        type: "string"
                        description: ""
                    description: "数据"
    put:
      summary: "cleaningTemplatesTemplateIdPut"
      description: "cleaningTemplatesTemplateIdPut"
      parameters:
      - name: "templateId"
        in: "path"
        description: ""
        required: true
        schema:
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                id:
                  type: "string"
                  description: ""
                name:
                  type: "string"
                  description: ""
                description:
                  type: "string"
                  description: ""
                instance:
                  type: "array"
                  description: ""
                  items:
                    type: "object"
                    properties:
                      id:
                        type: "string"
                        description: ""
                      overrides:
                        type: "object"
                        properties:
                          key:
                            type: "object"
                            properties: {}
                        description: ""
                        default: "new HashMap<>()"
                    description: "OperatorInstance"
                  default: "new ArrayList<>()"
              description: ""
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "string"
                    description: "状态码"
                  message:
                    type: "string"
                    description: "消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "string"
                        description: ""
                      name:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      instance:
                        type: "array"
                        description: ""
                        items:
                          type: "object"
                          properties:
                            id:
                              type: "string"
                              description: ""
                            name:
                              type: "string"
                              description: ""
                            description:
                              type: "string"
                              description: ""
                            version:
                              type: "string"
                              description: ""
                            inputs:
                              type: "string"
                              description: ""
                            outputs:
                              type: "string"
                              description: ""
                            runtime:
                              type: "string"
                              description: ""
                            settings:
                              type: "string"
                              description: ""
                            isStar:
                              type: "boolean"
                              description: ""
                            createdAt:
                              type: "string"
                              description: ""
                            updatedAt:
                              type: "string"
                              description: ""
                          description: "OperatorResponse"
                        default: "new ArrayList<>()"
                      createdAt:
                        type: "string"
                        description: ""
                      updatedAt:
                        type: "string"
                        description: ""
                    description: "数据"
    delete:
      summary: "cleaningTemplatesTemplateIdDelete"
      description: "cleaningTemplatesTemplateIdDelete"
      parameters:
      - name: "templateId"
        in: "path"
        description: ""
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "string"
                    description: "状态码"
                  message:
                    type: "string"
                    description: "消息"
                  data:
                    type: "object"
                    properties: {}
                    description: "数据"
...